Extract key building for encrypted PDFs into a dedicated class
