+++ lib/pdf/reader.rb
@@ -314,6 +314,7 @@ require 'pdf/reader/rectangle'
 require 'pdf/reader/reference'
 require 'pdf/reader/register_receiver'
 require 'pdf/reader/null_security_handler'
+require 'pdf/reader/standard_key_builder'
 require 'pdf/reader/standard_security_handler'
 require 'pdf/reader/standard_security_handler_v5'
 require 'pdf/reader/unimplemented_security_handler'
